feat: working on master
[ST-40](https://trikystudio.atlassian.net/browse/ST-40)
[ST-42](https://trikystudio.atlassian.net/browse/ST-42)
[ST-44](https://trikystudio.atlassian.net/browse/ST-44)
[ST-46](https://trikystudio.atlassian.net/browse/ST-46)
[ST-47](https://trikystudio.atlassian.net/browse/ST-47)
[ST-49](https://trikystudio.atlassian.net/browse/ST-49)
[ST-51](https://trikystudio.atlassian.net/browse/ST-51)
[ST-53](https://trikystudio.atlassian.net/browse/ST-53)
[ST-55](https://trikystudio.atlassian.net/browse/ST-55)

+++ src/entity/discount.rs
@@ -8,7 +8,7 @@ pub struct Discount {
     #[serde(rename = "_id", skip_serializing_if = "Option::is_none")]
     pub id: Option<ObjectId>,
     pub title: String,
-    pub amount: f32,
+    pub amount: f64,
     pub created_by: Option<ObjectId>,
     pub created_at: DateTime,
     pub updated_at: DateTime,
